server/requirements.txt
alembic
arrow
attrs
awsebcli
connexion
Flask-Migrate
flask-sqlalchemy
gevent
psycopg2-binary
pytest
pytest-env
pytest-flask
pytest-pycharm
pytest-xdist
black
